Added hawk

+++ src/main/java/dk/itu/sass/teame/boundary/FileResource.java
@@ -16,6 +16,8 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.Response;
@@ -27,7 +29,10 @@ import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
 
 import com.google.gson.JsonObject;
 
+import dk.itu.sass.teame.controller.AccountController;
 import dk.itu.sass.teame.controller.FileController;
+import dk.itu.sass.teame.controller.AuthProcessor;
+import dk.itu.sass.teame.entity.Account;
 import dk.itu.sass.teame.entity.File;
 import net.jalg.hawkj.Algorithm;
 import net.jalg.hawkj.AuthHeaderParsingException;
@@ -42,48 +47,13 @@ public class FileResource {
 
 	private final String FILE_LOCATION = "\\Irina\\ITU\\Sem_1\\Security\\wildfly-10.0.0.Final\\Pictures";
 
-
 	@Inject
 	FileController fc;
 
 	@GET
-	public Response getFile(@QueryParam("id") String id, @HeaderParam("Server-Authorization") String serverAuth) {
-		
-		AuthorizationHeader header = null;
-		try {
-			AuthorizationHeader header2 = AuthorizationHeader.authorization("Hawk");
-			
-			String method = "POST";
-			String path = "login";
-			String host = "localhost";
-			int port = 8080;
-			
-			String userid = "1";
-			String key = "admin";
-			Algorithm algorithm = Algorithm.SHA_256;
-			
-			HawkContext hawk = HawkContext.request("POST", "login",
-			                                       "localhost", 8080)
-			                     .credentials(userid, "tis", algorithm)
-			                     .tsAndNonce(1460203791L, "A0612D128F09")
-			                     .hash(null).build();
-			
-			if(hawk.isValidMac(serverAuth)) {
-				System.out.println("VALID");
-			} else {
-				System.out.println("NOT VALID");
-			}
-			
-			header = hawk.createAuthorizationHeader();
-
-			
-			
-		} catch (AuthHeaderParsingException e1) {
-			e1.printStackTrace();
-		}	
-		
+	public Response getFile(@QueryParam("id") String id, @QueryParam("id") long accountId,
+			@Context ContainerRequestContext requestContext) {
 		JsonObject json = new JsonObject();
-		
 		Long fid = null;
 		try {
 			fid = Long.parseLong(id);
@@ -91,27 +61,25 @@ public class FileResource {
 			json.addProperty("Error", "Wrong file id: " + id);
 			return Response.status(Status.BAD_REQUEST).entity(json.toString()).build();
 		}
+		Account user = AccountController.getAccountById(accountId);
+		if (AuthProcessor.Authenticate(requestContext, user.getPassword(), user.getUsername())) {
+			File file = fc.getFile(fid);
 
-		File file = fc.getFile(fid);
-		
-		if(file==null)
-			Response.status(Status.INTERNAL_SERVER_ERROR).build();
-		
-		JsonObject jsonResponse = fileToJsonObject(file);
-		jsonResponse.addProperty("auth", serverAuth);
-		jsonResponse.addProperty("id", header.getId());
-		jsonResponse.addProperty("dlg", header.getDlg());
-		jsonResponse.addProperty("hash", header.getHash());
-		jsonResponse.addProperty("mac", header.getMac());
-		jsonResponse.addProperty("ts", header.getTs());
+			if (file == null)
+				Response.status(Status.INTERNAL_SERVER_ERROR).build();
 
-		return Response.ok().entity(jsonResponse.toString()).build();
+			JsonObject jsonResponse = fileToJsonObject(file);
+			return Response.ok().entity(jsonResponse.toString()).build();
+		} else {
+			return Response.status(Status.UNAUTHORIZED).build();
+		}
 	}
 
 	@POST
 	@Consumes(MediaType.MULTIPART_FORM_DATA)
 	@Produces(MediaType.APPLICATION_JSON)
-	public Response uploadFile(MultipartFormDataInput input, @QueryParam("userid") String userId) {
+	public Response uploadFile(MultipartFormDataInput input, @QueryParam("userid") String userId,
+			@Context ContainerRequestContext requestContext) {
 
 		JsonObject json = new JsonObject();
 
@@ -122,34 +90,39 @@ public class FileResource {
 			json.addProperty("Error", "Wrong user id: " + userId);
 			return Response.status(Status.BAD_REQUEST).entity(json.toString()).build();
 		}
+		Account user = AccountController.getAccountById(uid);
+		if (AuthProcessor.Authenticate(requestContext, user.getPassword(), user.getUsername())) {
 
-		Map<String, List<InputPart>> maps = input.getFormDataMap();
-		List<InputPart> f = maps.get("file");
+			Map<String, List<InputPart>> maps = input.getFormDataMap();
+			List<InputPart> f = maps.get("file");
 
-		MultivaluedMap<String, String> mv = f.get(0).getHeaders();
+			MultivaluedMap<String, String> mv = f.get(0).getHeaders();
 
-		// Hijacking filenames xD Timestamp/uuid to fix - but it's a cool
-		// feature.
-		String filename = getFileName(mv);
+			// Hijacking filenames xD Timestamp/uuid to fix - but it's a cool
+			// feature.
+			String filename = getFileName(mv);
 
-		java.nio.file.Path sti = null;
+			java.nio.file.Path sti = null;
 
-		try {
-			InputStream is = f.get(0).getBody(InputStream.class, null);
-			byte[] barr = IOUtils.toByteArray(is);
-			java.nio.file.Path p = Paths.get(FILE_LOCATION, filename);// #fail
-			Files.deleteIfExists(p); // #fail
-			Files.createDirectories(p.getParent());// #fail
-			sti = Files.write(p, barr, StandardOpenOption.CREATE_NEW);// #fail
-		} catch (Exception e) {
-			json.addProperty("Error", "Couldn't write file");
-			json.addProperty("Trace", e.getMessage());
-			return Response.status(Status.INTERNAL_SERVER_ERROR).entity(json.toString()).build();
+			try {
+				InputStream is = f.get(0).getBody(InputStream.class, null);
+				byte[] barr = IOUtils.toByteArray(is);
+				java.nio.file.Path p = Paths.get(FILE_LOCATION, filename);// #fail
+				Files.deleteIfExists(p); // #fail
+				Files.createDirectories(p.getParent());// #fail
+				sti = Files.write(p, barr, StandardOpenOption.CREATE_NEW);// #fail
+			} catch (Exception e) {
+				json.addProperty("Error", "Couldn't write file");
+				json.addProperty("Trace", e.getMessage());
+				return Response.status(Status.INTERNAL_SERVER_ERROR).entity(json.toString()).build();
+			}
+
+			File file = fc.uploadFile(uid, sti);
+
+			return Response.created(file.getPath().toUri()).entity(fileToJsonObject(file).toString()).build();
+		} else {
+			return Response.status(Response.Status.UNAUTHORIZED).build();
 		}
-
-		File file = fc.uploadFile(uid, sti);
-
-		return Response.created(file.getPath().toUri()).entity(fileToJsonObject(file).toString()).build();
 	}
 
 	private String getFileName(MultivaluedMap<String, String> header) {
@@ -167,7 +140,7 @@ public class FileResource {
 		}
 		return "unknown";
 	}
-	
+
 	private JsonObject fileToJsonObject(File f) {
 		JsonObject json = new JsonObject();
 		json.addProperty("id", f.getId());
@@ -176,41 +149,48 @@ public class FileResource {
 		json.addProperty("timestamp", f.getTimestamp().toString());
 		return json;
 	}
-	
+
 	@GET
 	@Path("getallimages")
 	@Produces(MediaType.APPLICATION_JSON)
-	public Response getAllImagesFromUser(@QueryParam("id") String id){
-		
-		if(id.isEmpty() || id == null) {
+	public Response getAllImagesFromUser(@QueryParam("id") String id) {
+
+		if (id.isEmpty() || id == null) {
 			Response.status(Response.Status.BAD_REQUEST).build();
 		}
-		
+
 		String json = fc.getFilesByUser(Long.parseLong(id));
-		return Response.status(Response.Status.ACCEPTED).entity(json).build();		
+		return Response.status(Response.Status.ACCEPTED).entity(json).build();
 	}
-	
+
 	@POST
 	@Path("shareimage")
-	public Response shareImage(@FormParam("imageId") String imageId, 
-							   @FormParam("author") String authorId, 
-							   @FormParam("victim") String shareWithId){
-		
+	public Response shareImage(@FormParam("imageId") String imageId, @FormParam("author") String authorId,
+			@FormParam("victim") String shareWithId,
+			@Context ContainerRequestContext requestContext) {
+
 		JsonObject o = new JsonObject();
 		o.addProperty("author", authorId);
 		o.addProperty("sharedwith", shareWithId);
 		o.addProperty("imageid", imageId);
-		
-		if((imageId.isEmpty()  || imageId == null ) || 
-		   (authorId.isEmpty() || authorId == null ) || 
-		   (shareWithId.isEmpty() || shareWithId == null) ){
+
+		if ((imageId.isEmpty() || imageId == null) || (authorId.isEmpty() || authorId == null)
+				|| (shareWithId.isEmpty() || shareWithId == null)) {
 			Response.status(Response.Status.BAD_REQUEST).build();
 		}
-		
-		boolean b = fc.shareImage(Long.parseLong(imageId), Long.parseLong(authorId), Long.parseLong(shareWithId));
-		
-		if(b) return Response.status(Response.Status.ACCEPTED).entity(o.toString()).build();
-		else return Response.status(Response.Status.BAD_REQUEST).build();
+		long accountId = Long.parseLong(authorId);
+		Account acc = AccountController.getAccountById(accountId);
+		if (AuthProcessor.Authenticate(requestContext, acc.getPassword(), acc.getUsername())) {
+		boolean b = fc.shareImage(Long.parseLong(imageId), accountId, Long.parseLong(shareWithId));
+
+		if (b)
+			return Response.status(Response.Status.ACCEPTED).entity(o.toString()).build();
+		else
+			return Response.status(Response.Status.BAD_REQUEST).build();
+		}
+		else {
+			return Response.status(Response.Status.UNAUTHORIZED).build();
+		}
 	}
 
 }
